requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.52.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.17
reportlab==4.2.5
Jinja2==3.1.4
email-validator==2.1.0
httpx==0.27.0
supabase==2.3.4
